requirements.txt

requests
pysocks
